Hg19 genomic information of cell-type-specific marker CpGs inferred from DNA methylation atlas published by Moss 2018 Nat. Commun. (https://www.nature.com/articles/s41467-018-07466-6).

Default reference cell type-specific markers used as default in decemedip. This dataset represents a GRanges object that contains the collection of Illumina HumanMethylation450K probes that have methylation level less than 0.1 or greater than 0.9 in all tissue present in the atlas. Data source of the methylation atlas is from the MethAtlas GitHub repo (https://github.com/nloyfer/meth_atlas). For details of how the marker CpGs are selected, please refer to the decemedip manuscript.

Usage

data(hg19.ref.cts.se)

Format

An object of class GRanges. rowData(hg19.ref.cts.se) contains information of the selected probes.

Details

All coordinates are in hg19.

References

Moss, J., Magenheim, J., Neiman, D. et al. Comprehensive human cell-type methylation atlas reveals origins of circulating cell-free DNA in health and disease. Nat Commun 9, 5068 (2018). https://doi.org/10.1038/s41467-018-07466-6

Examples

  data(hg19.ref.cts.se)
  hg19.ref.cts.se
#> class: RangedSummarizedExperiment 
#> dim: 2500 25 
#> metadata(0):
#> assays(1): beta
#> rownames(2500): cg18856478 cg20820767 ... cg01071459 cg20726993
#> rowData names(4): probe label pos n_cpgs_100bp
#> colnames(25): Monocytes_EPIC B-cells_EPIC ... Upper_GI Uterus_cervix
#> colData names(0):